Backend Finalizado grazadeus

+++ Cadastro_Sementes_do_Amanha/app/models/attendance.py
@@ -3,8 +3,8 @@ from app.database import db
 class StudentAttendance(db.Model):
     __tablename__ = "student_attendance"
 
-    id = db.Column(db.Integer, primary_key=True)
-    student_id = db.Column(db.Integer, db.ForeignKey("students.id", ondelete="CASCADE"), nullable=False, index=True)
+    id = db.Column(db.BigInteger, primary_key=True)
+    student_id = db.Column(db.BigInteger, db.ForeignKey("students.id", ondelete="CASCADE"), nullable=False, index=True)
     data = db.Column(db.Date, nullable=False, index=True)
     status = db.Column(db.Text, nullable=False)
